Split travel composer checkout runtime
