+++ templates/operator/src/api/travel-composer-runtime.ts
@@ -13,21 +13,8 @@ import {
 import type { PricingBasis } from "@voyantjs/catalog/snapshot/schema"
 import type { EventBus } from "@voyantjs/core"
 import type { AnyDrizzleDb } from "@voyantjs/db"
-import { buildPaymentLinkUrl, financeService, type PaymentCompletedEvent } from "@voyantjs/finance"
-import type {
-  AncillarySelection,
-  FlightBookRequest,
-  FlightOffer,
-  FlightPassenger,
-  PassengerType,
-} from "@voyantjs/flights/contract/types"
+import type { PaymentCompletedEvent } from "@voyantjs/finance"
 import type { HonoBundle } from "@voyantjs/hono/plugin"
-import { createDemoFlightAdapter } from "@voyantjs/plugin-flights-demo"
-import {
-  NETOPIA_RUNTIME_CONTAINER_KEY,
-  netopiaService,
-  type ResolvedNetopiaRuntimeOptions,
-} from "@voyantjs/plugin-netopia"
 import { createCatalogPromotionEvaluator } from "@voyantjs/promotions/service-catalog-evaluator"
 import {
   type CancelComponentInput,
@@ -42,14 +29,10 @@ import {
   type ReleaseReservedComponentInput,
   type ReleaseReservedComponentResult,
   type ReserveComponentInput,
-  type ReserveComponentPreflightResult,
   type ReserveComponentResult,
   type ReserveTripDeps,
   type StartCheckoutDeps,
   type TravelComposerRoutesOptions,
-  type Trip,
-  type TripCheckoutInput,
-  type TripCheckoutResult,
   toBookingDraftV1,
   travelComposerService,
 } from "@voyantjs/travel-composer"
@@ -67,6 +50,11 @@ import {
   getOwnedBookingHandlerRegistryFromContext,
 } from "./lib/booking-engine-runtime"
 import { withDbFromEnv } from "./lib/db"
+import {
+  reserveNonCatalogComponent,
+  validateNonCatalogComponentBeforeReserve,
+} from "./travel-composer-flight-runtime"
+import { startTripCheckout } from "./travel-composer-trip-checkout"
 
 export function createOperatorTravelComposerRoutesOptions(): TravelComposerRoutesOptions {
   return {
@@ -260,109 +248,6 @@ async function releaseReservedComponent(
   }
 }
 
-async function validateNonCatalogComponentBeforeReserve(
-  c: Context,
-  input: ReserveComponentInput,
-): Promise<ReserveComponentPreflightResult | null> {
-  if (input.component.kind !== "flight_placeholder") return null
-
-  const flightDraft = asRecord(input.component.metadata)?.flightDraft
-  const draftRecord = asRecord(flightDraft)
-  const selectedOffer = draftRecord?.selectedOffer as FlightOffer | undefined
-  const offerId = stringValue(draftRecord?.offerId) ?? selectedOffer?.offerId
-  if (!selectedOffer || !offerId) {
-    return { status: "unavailable", reason: "flight_offer_required" }
-  }
-
-  try {
-    const priced = await getFlightAdapter(c).priceOffer(buildFlightAdapterContext(c), {
-      offerId,
-      offer: selectedOffer,
-    })
-    if (!priced.valid) {
-      return {
-        status: "unavailable",
-        reason: priced.invalidReason ?? "flight_offer_unavailable",
-        details: { offerId },
-      }
-    }
-
-    const previousPricing = asRecord(draftRecord?.pricing)
-    const ancillaryAmountCents = numberValue(previousPricing?.ancillaryAmountCents) ?? 0
-    const currentTotalAmountCents =
-      moneyToCents(priced.offer.totalPrice.amount) + ancillaryAmountCents
-    const previousTotalAmountCents =
-      input.component.componentTotalAmountCents ??
-      numberValue(previousPricing?.totalAmountCents) ??
-      0
-    const currentCurrency = priced.offer.totalPrice.currency
-    const previousCurrency =
-      input.component.componentCurrency ?? stringValue(previousPricing?.currency) ?? currentCurrency
-
-    if (
-      currentCurrency !== previousCurrency ||
-      currentTotalAmountCents !== previousTotalAmountCents
-    ) {
-      return {
-        status: "price_changed",
-        reason: "flight_price_changed",
-        details: {
-          offerId,
-          previous: {
-            currency: previousCurrency,
-            totalAmountCents: previousTotalAmountCents,
-          },
-          current: {
-            currency: currentCurrency,
-            totalAmountCents: currentTotalAmountCents,
-          },
-        },
-      }
-    }
-
-    return { status: "ok" }
-  } catch (error) {
-    return {
-      status: isExpiredOffer(selectedOffer) ? "expired" : "unavailable",
-      reason: error instanceof Error ? error.message : "flight_offer_unavailable",
-      details: { offerId },
-    }
-  }
-}
-
-async function reserveNonCatalogComponent(
-  c: Context,
-  input: ReserveComponentInput,
-): Promise<ReserveComponentResult | null> {
-  if (input.component.kind !== "flight_placeholder") return null
-
-  const flightDraft = asRecord(input.component.metadata)?.flightDraft
-  const selectedOffer = asRecord(flightDraft)?.selectedOffer as FlightOffer | undefined
-  if (!selectedOffer?.offerId) {
-    throw new Error("flight_offer_required")
-  }
-
-  const request: FlightBookRequest = {
-    offerId: selectedOffer.offerId,
-    offer: selectedOffer,
-    passengers: flightPassengersFromTravelerParty(input.envelope.travelerParty),
-    contact: flightContactFromTravelerParty(input.envelope.travelerParty),
-    paymentIntent: { type: "hold" },
-    ancillaries: asRecord(flightDraft)?.ancillaries as AncillarySelection | undefined,
-  }
-  const response = await getFlightAdapter(c).bookFlight(buildFlightAdapterContext(c), request)
-  const order = response.order
-
-  return {
-    status: order.status === "ticketed" ? "booked" : "held",
-    orderId: order.orderId,
-    providerRef: order.pnr ?? order.orderId,
-    supplierRef: order.orderId,
-    holdExpiresAt: order.paymentDeadline,
-    warnings: order.paymentDeadline ? undefined : ["flight_hold_deadline_missing"],
-  }
-}
-
 async function previewComponentCancellation(
   input: ComponentCancellationPreviewInput,
 ): Promise<ComponentCancellationPreview> {
@@ -473,86 +358,6 @@ async function startComponentCheckout(
   }
 }
 
-async function startTripCheckout(
-  c: Context,
-  input: TripCheckoutInput,
-): Promise<TripCheckoutResult> {
-  const db = getDb(c) as Parameters<typeof financeService.createPaymentSession>[0]
-  const pricing = await checkoutPricingForTrip(c, input.trip, input.request)
-  if (pricing.totalAmountCents <= 0) {
-    throw new Error("trip_checkout_total_required")
-  }
-
-  const billing = readBilling(input.trip.envelope.travelerParty)
-  const payerName = formatBillingName(billing)
-  const payerEmail = billing.contact?.email ?? null
-  if (!payerName || !payerEmail) {
-    throw new Error("trip_checkout_billing_required")
-  }
-  const paymentMethod = input.intent === "bank_transfer" ? "bank_transfer" : "credit_card"
-  const session = await financeService.createPaymentSession(db, {
-    targetType: "other",
-    targetId: input.trip.envelope.id,
-    idempotencyKey: `trip-checkout:${input.trip.envelope.id}:${pricing.currency}:${pricing.totalAmountCents}`,
-    clientReference: input.trip.envelope.id,
-    currency: pricing.currency,
-    amountCents: pricing.totalAmountCents,
-    status: "pending",
-    provider: input.intent === "bank_transfer" ? null : "netopia",
-    paymentMethod,
-    payerPersonId: billing.personId ?? null,
-    payerOrganizationId: billing.organizationId ?? null,
-    payerEmail,
-    payerName,
-    notes: buildTripPaymentSummary(input.trip, pricing.currency, pricing.allocations),
-    metadata: {
-      tripEnvelopeId: input.trip.envelope.id,
-      collectionCurrency: pricing.currency,
-      componentAllocations: pricing.allocations,
-      fxAllocations: pricing.allocations.filter((allocation) => allocation.fx),
-    },
-  })
-
-  if (input.intent !== "bank_transfer") {
-    try {
-      const runtime = getContainer(c)?.resolve(NETOPIA_RUNTIME_CONTAINER_KEY) as
-        | ResolvedNetopiaRuntimeOptions
-        | undefined
-      if (runtime) {
-        await netopiaService.startPaymentSession(
-          db,
-          session.id,
-          {
-            billing: synthesizeTripBilling(billing),
-            description: `Trip ${input.trip.envelope.id}`,
-          },
-          runtime,
-          undefined,
-        )
-      }
-    } catch (error) {
-      console.warn("[travel-composer] netopia start failed for trip payment session:", error)
-    }
-  }
-
-  return {
-    kind: input.intent === "bank_transfer" ? "bank_transfer_instructions" : "payment_session",
-    paymentSessionId: session.id,
-    checkoutUrl: buildPaymentLinkUrl(session.id, {
-      baseUrl: resolvePublicCheckoutBaseUrl(c.env as CloudflareBindings),
-    }),
-  }
-}
-
-function resolvePublicCheckoutBaseUrl(env: CloudflareBindings): string | null {
-  return (
-    env.PUBLIC_CHECKOUT_BASE_URL?.trim() ||
-    env.DASH_BASE_URL?.trim() ||
-    env.APP_URL?.trim().replace(/\/api\/?$/, "") ||
-    null
-  )
-}
-
 function checkoutResultToComponentResult(
   result: CatalogCheckoutStartResult,
 ): ComponentCheckoutResult {
@@ -661,382 +466,6 @@ function bookStatusToComponentStatus(status: BookEntityResult["status"]): "held"
   return status === "held" ? "held" : "booked"
 }
 
-function getFlightAdapter(c: Context) {
-  const baseUrl = (c.env as { FLIGHTS_DEMO_API_URL?: string }).FLIGHTS_DEMO_API_URL
-  if (!baseUrl) {
-    throw new Error(
-      "FLIGHTS_DEMO_API_URL is not set. Start `apps/flights-demo-api` and point this env at it.",
-    )
-  }
-  return createDemoFlightAdapter({ baseUrl })
-}
-
-function buildFlightAdapterContext(c: Context): { connectionId: string; correlationId?: string } {
-  return {
-    connectionId: "demo",
-    correlationId: c.req.header("x-request-id") ?? undefined,
-  }
-}
-
-interface TripCheckoutAllocation {
-  componentId: string
-  kind: string
-  bookingId: string | null
-  orderId: string | null
-  sourceCurrency: string
-  sourceAmountCents: number
-  targetCurrency: string
-  targetAmountCents: number
-  fx?: {
-    rate: number
-    provider: "voyant_data_fx"
-    quotedAt: string
-    validUntil?: string | null
-  }
-}
-
-async function checkoutPricingForTrip(
-  c: Context,
-  trip: Trip,
-  request: Record<string, unknown>,
-): Promise<{
-  currency: string
-  totalAmountCents: number
-  allocations: TripCheckoutAllocation[]
-}> {
-  const active = trip.components.filter(
-    (component) => component.status !== "removed" && component.status !== "cancelled",
-  )
-  const collectionCurrency =
-    stringValue(request.collectionCurrency) ?? trip.envelope.aggregateCurrency ?? "EUR"
-  const allocations: TripCheckoutAllocation[] = []
-
-  for (const component of active) {
-    const sourceCurrency = component.componentCurrency ?? collectionCurrency
-    const sourceAmountCents = component.componentTotalAmountCents ?? 0
-    if (sourceAmountCents <= 0) continue
-
-    if (sourceCurrency === collectionCurrency) {
-      allocations.push({
-        componentId: component.id,
-        kind: component.kind,
-        bookingId: component.bookingId,
-        orderId: component.orderId,
-        sourceCurrency,
-        sourceAmountCents,
-        targetCurrency: collectionCurrency,
-        targetAmountCents: sourceAmountCents,
-      })
-      continue
-    }
-
-    const fx = await quoteFx(c, sourceCurrency, collectionCurrency)
-    allocations.push({
-      componentId: component.id,
-      kind: component.kind,
-      bookingId: component.bookingId,
-      orderId: component.orderId,
-      sourceCurrency,
-      sourceAmountCents,
-      targetCurrency: collectionCurrency,
-      targetAmountCents: convertCents(sourceAmountCents, fx.rate),
-      fx: {
-        rate: fx.rate,
-        provider: "voyant_data_fx",
-        quotedAt: fx.quotedAt,
-        validUntil: fx.validUntil,
-      },
-    })
-  }
-
-  return {
-    currency: collectionCurrency,
-    totalAmountCents: allocations.reduce(
-      (sum, allocation) => sum + allocation.targetAmountCents,
-      0,
-    ),
-    allocations,
-  }
-}
-
-function buildTripPaymentSummary(
-  trip: Trip,
-  currency: string,
-  allocations?: TripCheckoutAllocation[],
-): string {
-  const lines = ["Trip payment summary"]
-  const byComponentId = new Map(
-    allocations?.map((allocation) => [allocation.componentId, allocation]),
-  )
-  for (const component of trip.components.filter(
-    (item) => item.status !== "removed" && item.status !== "cancelled",
-  )) {
-    const allocation = byComponentId.get(component.id)
-    if (allocation?.fx) {
-      lines.push(
-        `${componentDisplayName(component)} — ${formatCents(
-          allocation.sourceAmountCents,
-          allocation.sourceCurrency,
-        )} -> ${formatCents(allocation.targetAmountCents, allocation.targetCurrency)}`,
-      )
-    } else {
-      lines.push(
-        `${componentDisplayName(component)} — ${formatCents(
-          allocation?.targetAmountCents ?? component.componentTotalAmountCents,
-          allocation?.targetCurrency ?? component.componentCurrency ?? currency,
-        )}`,
-      )
-    }
-  }
-  const total = allocations?.reduce((sum, allocation) => sum + allocation.targetAmountCents, 0)
-  lines.push(
-    `Total payable — ${formatCents(total ?? trip.envelope.aggregateTotalAmountCents, currency)}`,
-  )
-  const fxAllocations = allocations?.filter((allocation) => allocation.fx) ?? []
-  if (fxAllocations.length > 0) {
-    lines.push("")
-    lines.push("FX rates")
-    for (const allocation of fxAllocations) {
-      lines.push(
-        `${allocation.sourceCurrency}->${allocation.targetCurrency}: ${allocation.fx?.rate} quoted ${allocation.fx?.quotedAt}`,
-      )
-    }
-  }
-  return lines.join("\n")
-}
-
-function componentDisplayName(component: Trip["components"][number]): string {
-  const metadata = asRecord(component.metadata)
-  const catalogItem = asRecord(metadata?.catalogItem)
-  const flightDraft = asRecord(metadata?.flightDraft)
-  const origin = stringValue(flightDraft?.origin)
-  const destination = stringValue(flightDraft?.destination)
-  if (origin && destination) return `${origin} -> ${destination}`
-  return (
-    stringValue(catalogItem?.name) ||
-    stringValue(component.title) ||
-    stringValue(component.description) ||
-    component.kind.replaceAll("_", " ")
-  )
-}
-
-function formatCents(amountCents: number | null | undefined, currency: string): string {
-  return ((amountCents ?? 0) / 100).toLocaleString("en-GB", {
-    style: "currency",
-    currency,
-  })
-}
-
-function moneyToCents(amount: string): number {
-  const parsed = Number.parseFloat(amount)
-  return Number.isFinite(parsed) ? Math.round(parsed * 100) : 0
-}
-
-function isExpiredOffer(offer: FlightOffer): boolean {
-  const expiresAt = offer.expiresAt ? new Date(offer.expiresAt) : null
-  return Boolean(expiresAt && expiresAt.getTime() <= Date.now())
-}
-
-interface FxQuote {
-  rate: number
-  quotedAt: string
-  validUntil?: string | null
-}
-
-async function quoteFx(
-  c: Context,
-  sourceCurrency: string,
-  targetCurrency: string,
-): Promise<FxQuote> {
-  if (sourceCurrency === targetCurrency) {
-    return { rate: 1, quotedAt: new Date().toISOString(), validUntil: null }
-  }
-
-  const env = c.env as { VOYANT_CLOUD_API_KEY?: string; VOYANT_CLOUD_API_URL?: string }
-  if (!env.VOYANT_CLOUD_API_KEY) {
-    throw new Error("trip_checkout_fx_requires_voyant_cloud_api_key")
-  }
-
-  const baseUrl = (env.VOYANT_CLOUD_API_URL ?? "https://api.voyantjs.com").replace(/\/$/, "")
-  const url = new URL(
-    `/data/fx/v1/fx/pair/${encodeURIComponent(sourceCurrency)}/${encodeURIComponent(
-      targetCurrency,
-    )}`,
-    `${baseUrl}/`,
-  )
-  const response = await fetch(url, {
-    headers: {
-      authorization: `Bearer ${env.VOYANT_CLOUD_API_KEY}`,
-      "x-voyant-sdk": "voyant-operator-travel-composer",
-    },
-  })
-  const payload = (await response.json().catch(() => null)) as Record<string, unknown> | null
-  if (!response.ok) {
-    throw new Error(`trip_checkout_fx_quote_failed:${response.status}`)
-  }
-
-  const body = asRecord(payload?.data) ?? payload
-  const rate = numberValue(body?.conversionRate) ?? numberValue(body?.conversion_rate)
-  if (!rate || rate <= 0) {
-    throw new Error("trip_checkout_fx_quote_invalid")
-  }
-
-  return {
-    rate,
-    quotedAt:
-      stringValue(body?.timeLastUpdateUtc) ??
-      stringValue(body?.time_last_update_utc) ??
-      unixSecondsToIso(
-        numberValue(body?.timeLastUpdateUnix) ?? numberValue(body?.time_last_update_unix),
-      ) ??
-      new Date().toISOString(),
-    validUntil:
-      stringValue(body?.timeNextUpdateUtc) ??
-      stringValue(body?.time_next_update_utc) ??
-      unixSecondsToIso(
-        numberValue(body?.timeNextUpdateUnix) ?? numberValue(body?.time_next_update_unix),
-      ),
-  }
-}
-
-function convertCents(amountCents: number, rate: number): number {
-  return Math.round(amountCents * rate)
-}
-
-function unixSecondsToIso(value: number | null): string | null {
-  if (!value || !Number.isFinite(value)) return null
-  return new Date(value * 1000).toISOString()
-}
-
-function flightPassengersFromTravelerParty(travelerParty: Record<string, unknown>) {
-  const travelers = Array.isArray(travelerParty.travelers)
-    ? (travelerParty.travelers as unknown[])
-    : []
-  const passengers = travelers.map((traveler, index) =>
-    flightPassengerFromTraveler(asRecord(traveler) ?? {}, index),
-  )
-  if (passengers.length > 0) return passengers
-
-  const billing = readBilling(travelerParty)
-  const names = splitName(formatBillingName(billing) ?? "Lead traveler")
-  return [
-    {
-      passengerId: "traveler_1",
-      type: "adult" as const,
-      firstName: names.firstName,
-      lastName: names.lastName,
-      dateOfBirth: "1990-01-01",
-      ...(billing.contact?.email ? { email: billing.contact.email } : {}),
-      ...(billing.contact?.phone ? { phone: billing.contact.phone } : {}),
-    },
-  ]
-}
-
-function flightPassengerFromTraveler(
-  traveler: Record<string, unknown>,
-  index: number,
-): FlightPassenger {
-  const type = passengerTypeFromRole(stringValue(traveler.role))
-  return {
-    passengerId:
-      stringValue(traveler.localId) || stringValue(traveler.personId) || `traveler_${index + 1}`,
-    type,
-    firstName: stringValue(traveler.firstName) || fallbackFirstName(type),
-    lastName: stringValue(traveler.lastName) || `${index + 1}`,
-    dateOfBirth: stringValue(traveler.dateOfBirth) || fallbackDobForPassengerType(type),
-    ...(stringValue(traveler.email) ? { email: stringValue(traveler.email) ?? undefined } : {}),
-    ...(stringValue(traveler.phone) ? { phone: stringValue(traveler.phone) ?? undefined } : {}),
-  }
-}
-
-function passengerTypeFromRole(role: string | null): PassengerType {
-  if (role === "child") return "child"
-  if (role === "infant") return "infant"
-  return "adult"
-}
-
-function fallbackFirstName(type: PassengerType): string {
-  if (type === "child") return "Child"
-  if (type === "infant") return "Infant"
-  return "Adult"
-}
-
-function fallbackDobForPassengerType(type: PassengerType): string {
-  if (type === "child") return "2016-01-01"
-  if (type === "infant") return "2025-01-01"
-  return "1990-01-01"
-}
-
-function flightContactFromTravelerParty(travelerParty: Record<string, unknown>) {
-  const billing = readBilling(travelerParty)
-  return {
-    ...(billing.contact?.email ? { email: billing.contact.email } : {}),
-    ...(billing.contact?.phone ? { phone: billing.contact.phone } : {}),
-  }
-}
-
-interface TripBillingInfo {
-  buyerType?: string | null
-  personId?: string | null
-  organizationId?: string | null
-  contact?: {
-    firstName?: string | null
-    lastName?: string | null
-    email?: string | null
-    phone?: string | null
-  }
-}
-
-function readBilling(travelerParty: Record<string, unknown>): TripBillingInfo {
-  const billing = asRecord(travelerParty.billing)
-  const contact = asRecord(billing?.contact)
-  return {
-    buyerType: stringValue(billing?.buyerType),
-    personId: stringValue(billing?.personId),
-    organizationId: stringValue(billing?.organizationId),
-    contact: {
-      firstName: stringValue(contact?.firstName),
-      lastName: stringValue(contact?.lastName),
-      email: stringValue(contact?.email),
-      phone: stringValue(contact?.phone),
-    },
-  }
-}
-
-function formatBillingName(billing: TripBillingInfo): string | null {
-  return [billing.contact?.firstName, billing.contact?.lastName]
-    .filter((part): part is string => Boolean(part))
-    .join(" ")
-    .trim()
-    ? [billing.contact?.firstName, billing.contact?.lastName]
-        .filter((part): part is string => Boolean(part))
-        .join(" ")
-    : null
-}
-
-function synthesizeTripBilling(billing: TripBillingInfo) {
-  const names = splitName(formatBillingName(billing) ?? "Trip customer")
-  return {
-    email: billing.contact?.email ?? "",
-    phone: billing.contact?.phone ?? "0000000000",
-    firstName: names.firstName,
-    lastName: names.lastName,
-    city: "TBD",
-    country: 642,
-    state: "TBD",
-    postalCode: "00000",
-    details: "Pending — customer to confirm at payment.",
-  }
-}
-
-function splitName(value: string): { firstName: string; lastName: string } {
-  const parts = value.trim().split(/\s+/).filter(Boolean)
-  return {
-    firstName: parts[0] ?? "Trip",
-    lastName: parts.slice(1).join(" ") || "Customer",
-  }
-}
-
 function componentReserveIdempotencyKey(componentId: string, quoteId: string): string {
   return `travel-composer:${componentId}:${quoteId}`.slice(0, 128)
 }
@@ -1090,10 +519,6 @@ function asRecord(value: unknown): Record<string, unknown> | undefined {
 
 function stringValue(value: unknown): string | null {
   return typeof value === "string" && value.length > 0 ? value : null
-}
-
-function numberValue(value: unknown): number | null {
-  return typeof value === "number" && Number.isFinite(value) ? value : null
 }
 
 function sumBookingDraftPax(value: unknown): number {
